Remove modals.d.ts
