+++ packages/desktop-client/src/components/modals/ConfirmCategoryDelete.tsx
@@ -4,13 +4,13 @@ import React, { useState } from 'react';
 import { type CategoryGroupEntity } from 'loot-core/src/types/models';
 
 import { theme } from '../../style';
-import { type CommonModalProps } from '../../types/modals';
 import { CategoryAutocomplete } from '../autocomplete/CategoryAutocomplete';
 import { Block } from '../common/Block';
 import { Button } from '../common/Button';
 import { Modal } from '../common/Modal';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 type ConfirmCategoryDeleteProps = {
   modalProps: CommonModalProps;
