+++ packages/desktop-client/src/components/modals/ConfirmTransactionEdit.tsx
@@ -1,11 +1,11 @@
 // @ts-strict-ignore
 import React from 'react';
 
-import { type CommonModalProps } from '../../types/modals';
 import { Block } from '../common/Block';
 import { Button } from '../common/Button';
 import { Modal } from '../common/Modal';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 type ConfirmTransactionEditProps = {
   modalProps: Partial<CommonModalProps>;
