+++ packages/desktop-client/src/components/modals/CreateEncryptionKey.tsx
@@ -8,7 +8,6 @@ import { getCreateKeyError } from 'loot-core/src/shared/errors';
 
 import { type BoundActions } from '../../hooks/useActions';
 import { theme } from '../../style';
-import { type CommonModalProps } from '../../types/modals';
 import { ButtonWithLoading } from '../common/Button';
 import { ExternalLink } from '../common/ExternalLink';
 import { InitialFocus } from '../common/InitialFocus';
@@ -17,6 +16,7 @@ import { Modal, ModalButtons } from '../common/Modal';
 import { Paragraph } from '../common/Paragraph';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 type CreateEncryptionKeyProps = {
   modalProps: CommonModalProps;
