+++ packages/desktop-client/src/components/modals/FixEncryptionKey.tsx
@@ -7,7 +7,6 @@ import { getTestKeyError } from 'loot-core/src/shared/errors';
 
 import { type BoundActions } from '../../hooks/useActions';
 import { theme } from '../../style';
-import { type CommonModalProps } from '../../types/modals';
 import { Button, ButtonWithLoading } from '../common/Button';
 import { ExternalLink } from '../common/ExternalLink';
 import { InitialFocus } from '../common/InitialFocus';
@@ -16,6 +15,7 @@ import { Modal, ModalButtons } from '../common/Modal';
 import { Paragraph } from '../common/Paragraph';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 type FixEncryptionKeyProps = {
   modalProps: CommonModalProps;
