+++ packages/desktop-client/src/components/modals/Notes.tsx
@@ -5,10 +5,10 @@ import { useLiveQuery } from 'loot-core/src/client/query-hooks';
 import { q } from 'loot-core/src/shared/query';
 
 import { SvgCheck } from '../../icons/v2';
-import { type CommonModalProps } from '../../types/modals';
 import { Button } from '../common/Button';
 import { Modal } from '../common/Modal';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 import { Notes as NotesComponent } from '../Notes';
 
 type NotesProps = {
