+++ packages/desktop-client/src/components/modals/PlaidExternalMsg.tsx
@@ -3,13 +3,13 @@ import React, { useState, useRef } from 'react';
 
 import { AnimatedLoading } from '../../icons/AnimatedLoading';
 import { theme } from '../../style';
-import { type CommonModalProps } from '../../types/modals';
 import { Error } from '../alerts';
 import { Button } from '../common/Button';
 import { Modal, ModalButtons } from '../common/Modal';
 import { Paragraph } from '../common/Paragraph';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 function renderError(error) {
   return (
