+++ packages/desktop-client/src/components/modals/ReportBudgetSummary.tsx
@@ -4,12 +4,12 @@ import { sheetForMonth } from 'loot-core/src/shared/months';
 import * as monthUtils from 'loot-core/src/shared/months';
 
 import { styles } from '../../style';
-import { type CommonModalProps } from '../../types/modals';
 import { ExpenseTotal } from '../budget/report/budgetsummary/ExpenseTotal';
 import { IncomeTotal } from '../budget/report/budgetsummary/IncomeTotal';
 import { Saved } from '../budget/report/budgetsummary/Saved';
 import { Modal } from '../common/Modal';
 import { Stack } from '../common/Stack';
+import { type CommonModalProps } from '../Modals';
 import { NamespaceContext } from '../spreadsheet/NamespaceContext';
 
 type ReportBudgetSummaryProps = {
