+++ packages/desktop-client/src/components/modals/SwitchBudgetType.tsx
@@ -2,12 +2,12 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { type CommonModalProps } from '../../types/modals';
 import { Button } from '../common/Button';
 import { ExternalLink } from '../common/ExternalLink';
 import { Modal } from '../common/Modal';
 import { Paragraph } from '../common/Paragraph';
 import { Text } from '../common/Text';
+import { type CommonModalProps } from '../Modals';
 
 type SwitchBudgetTypeProps = {
   modalProps: CommonModalProps;
