+++ packages/desktop-client/src/components/schedules/ScheduleLink.tsx
@@ -6,11 +6,11 @@ import { send } from 'loot-core/src/platform/client/fetch';
 import { type Query } from 'loot-core/src/shared/query';
 
 import { type BoundActions } from '../../hooks/useActions';
-import { type CommonModalProps } from '../../types/modals';
 import { Modal } from '../common/Modal';
 import { Search } from '../common/Search';
 import { Text } from '../common/Text';
 import { View } from '../common/View';
+import { type CommonModalProps } from '../Modals';
 
 import { ROW_HEIGHT, SchedulesTable } from './SchedulesTable';
